Fix race condition in FileStorage
There's a race condition when Bootscale tries to write serialized data
on disk when it's run in highly concurrent environment, for example, via
parallel_test command with big number of processes, or when many
workers starting in cold boot in production.

This commit brings `atomic_write` implementation from Sprockets
library to Bootscale. First, it writes to a temporary file with unique
name per thread, and renames file to destination when write is
completed.